requirements.txt
adal==1.2.1
aioredis==1.2.0
amqp==2.5.2
anyjson==0.3.3
asgi-redis==1.4.3
asgiref==3.1.2
asn1crypto==0.24.0
astroid==2.2.5
async-timeout==3.0.1
attrs==19.3.0
autobahn==19.3.3
Automat==0.7.0
Babel==2.7.0
backports.csv==1.0.7
bcrypt==3.1.5
billiard==3.6.1.0
cachetools==3.1.0
celery==4.3.0
celery-with-redis==3.0
certifi==2019.3.9
cffi==1.14.0
channels==2.2.0
channels-redis==2.4.0
chardet==3.0.4
Click==7.0
constantly==15.1.0
cryptography==2.4.2
cssselect==1.0.3
daphne==2.3.0
defusedxml==0.5.0
diff-match-patch==20181111
Django==2.1.4
django-auth-ldap==1.7.0
django-celery-results==1.1.2
django-cors-headers==3.0.2
django-crispy-forms==1.7.2
django-formtools==2.1
django-import-export==1.1.0
django-pure-pagination==0.3.0
django-ranged-response==0.2.0
django-simple-captcha==0.4.6
django-widget-tweaks==1.4.5
djangorestframework==3.10.2
djangorestframework-jwt==1.11.0
DjangoUeditor==1.8.143
dnspython==1.16.0
dwebsocket==0.5.10
et-xmlfile==1.0.1
eventlet==0.25.1
flower==0.9.3
future==0.17.1
gevent==1.4.0
gevent-websocket==0.10.1
google-auth==1.6.3
greenlet==0.4.15
hiredis==1.0.0
httplib2==0.12.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.1.0
incremental==17.5.0
isort==4.3.15
itsdangerous==1.1.0
jdcal==1.4
Jinja2==2.10.1
kafka-python==1.4.7
kombu==4.6.7
kubernetes==11.0.0
lazy-object-proxy==1.3.1
lxml==4.3.3
MarkupSafe==1.1.1
mccabe==0.6.1
monotonic==1.5
more-itertools==8.0.0
msgpack==0.6.1
msgpack-python==0.5.6
multi-key-dict==2.0.3
mysql-connector-python==8.0.15
oauthlib==3.0.1
odfpy==1.4.0
openpyxl==2.5.12
paramiko==2.4.2
parsel==1.5.1
pbr==5.4.3
Pillow==7.2.0
protobuf==3.7.1
pyasn1==0.4.4
pyasn1-modules==0.2.4
pycparser==2.19
pycrypto==2.6.1
PyDispatcher==2.0.5
PyHamcrest==1.9.0
PyJWT==1.7.1
pylint==2.3.1
PyMySQL==0.9.3
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyotp==2.3.0
pypng==0.0.18
PyQRCode==1.2.1
python-dateutil==2.8.0
python-gitlab==1.12.1
python-jenkins==1.5.0
python-ldap==3.2.0
pytz==2018.7
PyYAML==3.13
queuelib==1.5.0
rc4-python3==1.1.3
redis==3.3.11
requests==2.21.0
requests-oauthlib==1.2.0
rsa==4.0
scipy==1.5.1
service-identity==18.1.0
six==1.12.0
sqlparse==0.3.0
tablib==0.12.1
txaio==18.8.1
typed-ast==1.4.1
unicodecsv==0.14.1
urllib3==1.25.9
vine==1.3.0
w3lib==1.20.0
websocket-client==0.56.0
Werkzeug==0.14.1
wrapt==1.11.1
xlrd==1.2.0
xlwt==1.3.0
zipp==3.1.0
zope.interface==5.1.0
